# openapi-spec.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/sports/leagues/{sportSlug}": {
            "get": {
                "operationId": "SportsController_getLeaguesBySport",
                "summary": "Get leagues for a specific sport",
                "parameters": [
                    {
                        "name": "sportSlug",
                        "required": true,
                        "in": "path",
                        "description": "Sport slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Sports"
                ]
            }
        },
        "/api/v1/sports/{slug}/matches": {
            "get": {
                "operationId": "SportsController_getMatchesBySport",
                "summary": "Get matches by sport slug",
                "parameters": [
                    {
                        "name": "slug",
                        "required": true,
                        "in": "path",
                        "description": "Sport slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "enum": [
                                "upcoming",
                                "live",
                                "finished"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Sports"
                ]
            }
        },
        "/api/v1/sports/matches": {
            "get": {
                "operationId": "SportsController_getMatches",
                "summary": "Get matches with filters",
                "parameters": [
                    {
                        "name": "featured",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "dateTo",
                        "required": false,
                        "in": "query",
                        "description": "End date filter (ISO)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dateFrom",
                        "required": false,
                        "in": "query",
                        "description": "Start date filter (ISO)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "enum": [
                                "upcoming",
                                "live",
                                "finished"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "name": "leagueSlug",
                        "required": false,
                        "in": "query",
                        "description": "Filter by league",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sportSlug",
                        "required": false,
                        "in": "query",
                        "description": "Filter by sport",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Sports"
                ]
            }
        },
...
        "/api/v1/sports/league/{slug}/matches": {
            "get": {
                "operationId": "SportsController_getMatchesByLeague",
                "summary": "Get matches by league slug",
                "parameters": [
                    {
                        "name": "slug",
                        "required": true,
                        "in": "path",
                        "description": "League slug",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "enum": [
                                "upcoming",
                                "live",
                                "finished"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Sports"
                ]
            }
        },
...
        "/api/v1/sports/matches/{id}": {
            "get": {
                "operationId": "SportsController_getMatchById",
                "summary": "Get match details with all markets and odds",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Match ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Sports"
                ]
            }
        },
        "/api/v1/sports/bets": {
            "post": {
                "operationId": "SportsController_placeBet",
                "summary": "Place a bet (single or combo)",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "single",
                                            "combo"
                                        ],
                                        "description": "Bet type"
                                    },
                                    "stake": {
                                        "type": "number",
                                        "description": "Stake amount"
                                    },
                                    "coinType": {
                                        "type": "string",
                                        "enum": [
                                            "GC",
                                            "SC"
                                        ],
                                        "description": "Coin type to bet with"
                                    },
                                    "selections": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "matchId": {
                                                    "type": "string"
                                                },
                                                "marketId": {
                                                    "type": "string"
                                                },
                                                "oddId": {
                                                    "type": "string"
                                                },
                                                "selection": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "matchId",
                                                "marketId",
                                                "oddId",
                                                "selection"
                                            ]
                                        },
                                        "description": "Bet selections"
                                    }
                                },
                                "required": [
                                    "type",
                                    "stake",
                                    "coinType",
                                    "selections"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Sports"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            },
            "get": {
                "operationId": "SportsController_getUserBets",
                "summary": "Get user bet history",
                "parameters": [
                    {
                        "name": "dateTo",
                        "required": false,
                        "in": "query",
                        "description": "End date filter (ISO)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dateFrom",
                        "required": false,
                        "in": "query",
                        "description": "Start date filter (ISO)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "enum": [
                                "single",
                                "combo"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "enum": [
                                "pending",
                                "won",
                                "lost",
                                "void",
                                "cashout"
                            ],
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Sports"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/sports/bets/{betId}": {
            "get": {
                "operationId": "SportsController_getBetById",
                "summary": "Get specific bet details",
                "parameters": [
                    {
                        "name": "betId",
                        "required": true,
                        "in": "path",
                        "description": "Bet ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Sports"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/sports/bets/{betId}/cashout": {
            "post": {
                "operationId": "SportsController_cashoutBet",
                "summary": "Cashout a pending bet",
                "parameters": [
                    {
                        "name": "betId",
                        "required": true,
                        "in": "path",
                        "description": "Bet ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Sports"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "WBC 2026 API",
        "description": "WBC 2026 Social Casino Platform API Documentation",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "JWT-auth": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "name": "JWT",
                "description": "Enter JWT token",
                "in": "header"
            }
        },
...
    }
}